feat: working canvas
This is the upload of my work for the Pixel Art Maker project. 

It is the last project in the Introduction to JavaScript on Udacity's Introduction to Programming Nanodegree.
I did not make any changes to the HTML file but changes were made to the styles.css and especially to the designs.js file.
For the CSS, I just added some effects when hovering over the canvas and changed the color of the <h1> to the same color selected for painting. The JavaScript file was written respecting the specifications.